Add files via upload
The Detec Behavior for Airport Passsenger project was developed in Python using VS Code, training the model, which is a part dedicated to detecting suspicious behavior of humans, while there are two parts in the project, one dedicated to detecting faces of those wanted by the state and a part that includes the project’s API, the admin’s dashboard and a mobile application for the field security officer at the airport. It was completed in 2025.